FORK when no fast rebuild
this is corner case:
if you are slow rebuilding over last 24hours and somehow when you have restarted you are the only node that has orphan block, the rebuild is looping forever to find a compatible chain.

Ideally we should improve the state machine to check if the PBFT majority is on another chain.